[twone_jin] 1. 컬럼 추가 후 'Done' 컬럼 checkmark 이미지 사라지는 오류 수정 2. 사용자 관리 페이지로 이동 시 사이드바에 '보드 리스트' 출력되도록 수정 230119/20:05

+++ src/main/java/com/shinjin/twone/controller/BoardController.java
@@ -122,7 +122,6 @@ public class BoardController {
         return result;
     }
 
-
     /*** 컬럼 생성 ***/
     @GetMapping("/project/addcolumn")
     @ResponseBody
@@ -182,11 +181,15 @@ public class BoardController {
             result += "<div id=" + cdto.getColSeq() + " class=\"col-lg-2 col-md-6\" style=\"min-width: 300px\">";
             result += "<div class=\"card info-card sales-card\">";
             result += "<div class=\"filter\">";
-            result += "<a class=\"icon\" href=\"#\" data-bs-toggle=\"dropdown\"><i class=\"bi bi-three-dots\"></i></a>";
-            result += "<ul class=\"dropdown-menu dropdown-menu-end dropdown-menu-arrow\">";
-            result += "<li><a class=\"dropdown-item\" href=\"#\">이슈 추가</a></li>";
-            result += "<li><a class=\"dropdown-item\" href=\"javascript:deletecolumn(" + cdto.getColSeq() + ")\">컬럼 삭제</a></li>";
-            result += "</ul>";
+            if(cdto.getColType() == 1){
+                result += "<a class=\"icon\" href=\"#\" data-bs-toggle=\"dropdown\"><img src=\"../resources/bootstrap/img/checkmark.png\" width=\"20\"></a>";
+            } else {
+                result += "<a class=\"icon\" href=\"#\" data-bs-toggle=\"dropdown\"><i class=\"bi bi-three-dots\"></i></a>";
+                result += "<ul class=\"dropdown-menu dropdown-menu-end dropdown-menu-arrow\">";
+                result += "<li><a class=\"dropdown-item\" href=\"#\">이슈 추가</a></li>";
+                result += "<li><a class=\"dropdown-item\" href=\"javascript:deletecolumn(" + cdto.getColSeq() + ")\">컬럼 삭제</a></li>";
+                result += "</ul>";
+            }
             result += "</div>";
             result += "<div class=\"card-body\">";
             result += "<h5 class=\"card-title\">" + cdto.getColName() + "</h5>";
